feat: extract AI into standalone aiservice; frontgateservice calls aiservice directly
