+++ Makefile
@@ -111,7 +111,12 @@ build-contrastdatacollector:
 	cd services/contrastdatacollector && \
 	docker build -t contrastdatacollector:latest  .
 
-build-containers: build-dataservice build-webhookservice build-frontgateservice build-console-ui build-exploit-server build-imageservice build-labelservice build-docservice build-reportservice build-contrastdatacollector
+build-aiservice:
+	@echo "Building aiservice..."
+	cd services/aiservice && \
+	docker build -t aiservice:latest  .
+
+build-containers: build-dataservice build-webhookservice build-frontgateservice build-console-ui build-exploit-server build-imageservice build-labelservice build-docservice build-reportservice build-contrastdatacollector build-aiservice
 	@echo "\nBuilding containers complete."
 
 run-helm: build-containers
