+++ services/aiservice/src/main/java/com/contrast/aiservice/AiServiceApplication.java
@@ -1,0 +1,11 @@
+package com.contrast.aiservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AiServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(AiServiceApplication.class, args);
+    }
+}
