+++ services/aiservice/src/main/resources/application.properties
@@ -1,0 +1,6 @@
+server.port=8080
+spring.application.name=aiservice
+
+ai.demo.enabled=true
+ai.demo.openai.base-url=http://ollama:11434/v1
+ai.demo.model=smollm2:135m
