+++ services/dataservice/src/main/resources/application.properties
@@ -28,11 +28,6 @@ spring.jpa.properties.hibernate.format_sql=true
 spring.datasource.hikari.connection-timeout=60000
 spring.datasource.hikari.maximum-pool-size=5
 
-# AI Demo Service Configuration (Ollama local model)
-ai.demo.enabled=true
-ai.demo.openai.base-url=http://ollama:11434/v1
-ai.demo.model=smollm2:135m
-
 # Spring Boot Actuator configuration
 management.endpoints.web.exposure.include=health,info,metrics,env,beans,configprops,heapdump,shutdown
 management.endpoint.health.show-details=always
